refactring

+++ Calculator/AddOperator.cs
@@ -4,7 +4,7 @@ namespace Calculator
 {
     public class AddOperator : IOperator
     {
-        public BigInteger doOperation(BigInteger num1, BigInteger num2)
+        public decimal doOperation(decimal num1, decimal num2)
         {
             return num1 + num2;
         }
