+++ Calculator/DivideOperator.cs
@@ -5,7 +5,7 @@ namespace Calculator
 {
     public class DivideOperator : IOperator
     {
-        public BigInteger doOperation(BigInteger num1, BigInteger num2)
+        public decimal doOperation(decimal num1, decimal num2)
         {
             if (num2 == 0)
             {
