+++ Calculator/IOperator.cs
@@ -4,6 +4,6 @@ namespace Calculator
 {
     public interface IOperator
     {
-        public BigInteger doOperation(BigInteger num1, BigInteger num2);
+        public decimal doOperation(decimal num1, decimal num2);
     }
 }
